backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
openpyxl
python-dotenv
google-generativeai